Support for optional parameters & default values.
Config parameters have a new field: default value.
If NULL, the parameter is mandatory, and must be set as before.
Otherwise, it contains a string with a default value
(with the same syntax as the value in config file would have).

Also added a field with parameter description aka modinfo(8).
This is not used yet, but should be a source for future online
help, showing available filters and their parameters.

Next, aded proper error codes for missing config file sections,
so that no Invalid error is shown anymore.

Also improved some logging messages.